initial quest api and ui setup

+++ pages/api/s3-upload/index.ts
@@ -12,11 +12,19 @@ export default APIRoute.configure({
     const roomId: string = req.body.roomId
 
     // delete old image if exists
-    await prisma.s3Image.delete({
+    const exists = await prisma.s3Image.findFirst({
       where: {
         roomId,
       },
     })
+
+    if (exists) {
+      await prisma.s3Image.delete({
+        where: {
+          roomId,
+        },
+      })
+    }
 
     const url = `images/${roomId}/${filename}`
